Moved NFC reader to InformationActivity
