+++ app/src/main/java/com/rezoleo/zer0data/InformationActivity.java
@@ -1,17 +1,108 @@
 package com.rezoleo.zer0data;
 
+import android.app.PendingIntent;
+import android.content.Intent;
+import android.content.IntentFilter;
+import android.nfc.NfcAdapter;
+import android.nfc.Tag;
+import android.nfc.tech.MifareClassic;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 import android.view.View;
+import android.widget.TextView;
 
 public class InformationActivity extends AppCompatActivity {
+
+    NfcAdapter mAdapter;
+    PendingIntent mPendingIntent;
+    IntentFilter[] mFilters;
+    String[][] mTechLists;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_information);
 
-        hideView(findViewById(R.id.card_recognized));
+//        hideView(findViewById(R.id.card_recognized));
+
+        mAdapter = NfcAdapter.getDefaultAdapter(this);
+
+        mPendingIntent = PendingIntent.getActivity(this, 0,
+                new Intent(this, getClass()).addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP), 0);
+
+        IntentFilter ndef = new IntentFilter(NfcAdapter.ACTION_TECH_DISCOVERED);
+
+        try {
+            ndef.addDataType("*/*");
+        } catch (IntentFilter.MalformedMimeTypeException e) {
+            throw new RuntimeException("fail", e);
+        }
+        mFilters = new IntentFilter[] {
+                ndef,
+        };
+
+        // Setup a tech list for all NfcF tags
+        mTechLists = new String[][] { new String[] { MifareClassic.class.getName() } };
+
+        Intent intent = getIntent();
+
+        resolveIntent(intent);
+    }
+
+
+    private void resolveIntent(Intent intent) {
+        // 1) Parse the intent and get the action that triggered this intent
+        String action = intent.getAction();
+
+        // 2) Check if it was triggered by a tag discovered interruption.
+        if (NfcAdapter.ACTION_TECH_DISCOVERED.equals(action)) {
+
+            //  3) Get an instance of the TAG from the NfcAdapter
+            Tag tagFromIntent = intent.getParcelableExtra(NfcAdapter.EXTRA_TAG);
+
+            String tagUid = ByteArrayToHexString(tagFromIntent.getId());
+            Log.i("NFC", tagUid);
+
+            TextView tv = (TextView) findViewById(R.id.card_recognized);
+            tv.setText(tagUid);
+        }
+    }
+
+
+    // Converting byte[] to hex string:
+    private String ByteArrayToHexString(byte[] inArray) {
+        int i, j, in;
+        String[] hex = {"0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "A", "B", "C", "D", "E", "F"};
+        String out = "";
+        for (j = 0; j < inArray.length; ++j) {
+            in = (int) inArray[j] & 0xff;
+            i = (in >> 4) & 0x0f;
+            out += hex[i];
+            i = in & 0x0f;
+            out += hex[i];
+        }
+        return out;
+    }
+
+    @Override
+    protected void onNewIntent(Intent intent) {
+        super.onNewIntent(intent);
+        Log.i("Foreground dispatch", "Discovered tag with intent: " + intent);
+        resolveIntent(intent);
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        mAdapter.enableForegroundDispatch(this, mPendingIntent, mFilters, mTechLists);
+    }
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+        mAdapter.disableForegroundDispatch(this);
     }
 
     protected void hideView(View viewToHide)
